+++ packages/common/esbuild/src/plugins/index.ts
@@ -1,2 +1,0 @@
-export * from "./bundle-size-logger-plugin"
-export * from "./esbuild-decorator-plugin"