+++ packages/common/mdx-vite/src/ai-knowledge/generator/knowledge-generator.ts
@@ -33,6 +33,7 @@ import type {AiKnowledgeConfig} from "../types"
 import {getConfig, setConfig} from "./config"
 import {formatDemos} from "./demo-plugin"
 import {PropFormatter} from "./doc-props-plugin"
+import {filterTextDirectives} from "./filter-text-directives"
 import type {MdxFlowExpression, ProcessedPage} from "./generator.types"
 import {formatNpmInstallTabs} from "./npm-install-tabs-plugin"
 import {formatThemeNodes} from "./qds-theme-plugin"
@@ -306,6 +307,7 @@ export class KnowledgeGenerator {
       .use(this.formatFrontmatterExpressions(frontmatter))
       .use(await formatThemeNodes())
       .use(formatDemos(pageInfo.demosFolder))
+      .use(filterTextDirectives)
       .use(this.transformRelativeUrls(pageInfo.url))
 
     this.applyPlugins(pageInfo, processor)